test: constructor unit tests

+++ test/shared/BaseTest.sol
@@ -25,7 +25,7 @@ contract BaseTest is Test, Deployers {
         uint256 numTokensToSell;
         uint256 startingTime;
         uint256 endingTime;
-        uint256 gamma;
+        int24 gamma;
         uint256 epochLength;
         uint24 fee;
         int24 tickSpacing;
@@ -36,7 +36,7 @@ contract BaseTest is Test, Deployers {
     uint256 constant DEFAULT_NUM_TOKENS_TO_SELL = 100_000e18;
     uint256 constant DEFAULT_STARTING_TIME = 1 days;
     uint256 constant DEFAULT_ENDING_TIME = 2 days;
-    uint256 constant DEFAULT_GAMMA = 800;
+    int24 constant DEFAULT_GAMMA = 800;
     uint256 constant DEFAULT_EPOCH_LENGTH = 400 seconds;
     // default to feeless case for now
     uint24 constant DEFAULT_FEE = 0;
